+++ src/Symfony/Component/Finder/CHANGELOG.md
@@ -1,6 +1,16 @@
 CHANGELOG
 =========
 
+4.0.0
+-----
+
+ * removed `ExceptionInterface`
+
+3.4.0
+-----
+
+ * deprecated `Symfony\Component\Finder\Iterator\FilterIterator`
+
 3.3.0
 -----