+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -3,11 +3,15 @@ CHANGELOG
 
 4.0.0
 -----
- 
+
  * removed `ControllerResolverInterface::getArguments()`
  * removed `TraceableControllerResolver::getArguments()`
  * removed `ControllerResolver::getArguments()` and the ability to resolve arguments
  * removed the `argument_resolver` service dependency from the `debug.controller_resolver`
+ * removed `LazyLoadingFragmentHandler::addRendererService()`
+ * removed `Psr6CacheClearer::addPool()`
+ * removed `Extension::addClassesToCompile()` and `Extension::getClassesToCompile()`
+ * removed `Kernel::loadClassCache()`, `Kernel::doLoadClassCache()` and `Kernel::setClassCache()`
 
 3.3.0
 -----
@@ -19,6 +23,7 @@ CHANGELOG
  * added the possibility to change the query string parameter used by `UriSigner`
  * deprecated `LazyLoadingFragmentHandler::addRendererService()`
  * deprecated `Extension::addClassesToCompile()`
+ * deprecated `Psr6CacheClearer::addPool()`
 
 3.2.0
 -----